+++ ayame.go
@@ -6,12 +6,13 @@ import (
 	"io/ioutil"
 	"log"
 	"net/http"
+	"time"
 
 	logrus "github.com/sirupsen/logrus"
 	"gopkg.in/yaml.v2"
 )
 
-var AyameVersion = "19.02.1"
+var AyameVersion = "19.08.0"
 
 type AyameOptions struct {
 	LogDir         string `yaml:"log_dir"`
@@ -20,7 +21,8 @@ type AyameOptions struct {
 	Addr           string `yaml:"addr"`
 	Port           int    `yaml:"port"`
 	OverWsPingPong bool   `yaml:"over_ws_ping_pong"`
-	AuthWebhookUrl string `yaml:"auth_webhook_url"`
+	AuthWebhookURL string `yaml:"auth_webhook_url"`
+	AllowOrigin    string `yaml:"allow_origin"`
 }
 
 var (
@@ -29,19 +31,17 @@ var (
 	logger  *logrus.Logger
 )
 
-// Initialize
+// initialization
 func init() {
-	configFilePath := flag.String("c", "./config.yaml", "ayame configuration file (yaml)")
-	// read yaml file
+	configFilePath := flag.String("c", "./config.yaml", "ayame configuration file path (yaml)")
+	// yaml file set
 	buf, err := ioutil.ReadFile(*configFilePath)
 	if err != nil {
-		// Fatal eit on read error
 		log.Fatal("cannot open config file, err=", err)
 	}
-	// parse yaml
+	// yaml parse
 	err = yaml.Unmarshal(buf, &Options)
 	if err != nil {
-		// Fatal exit on parse failure
 		log.Fatal("cannot parse config file, err=", err)
 	}
 }
@@ -49,10 +49,10 @@ func init() {
 func main() {
 	flag.Parse()
 	args := flag.Args()
-	// processing arguments
+	// argument processing
 	if len(args) > 0 {
 		if args[0] == "version" {
-			fmt.Printf("WebRTC Signaling Server Ayame version %s", AyameVersion)
+			fmt.Printf("WebRTC Signaling Server Ayame version %s\n", AyameVersion)
 			return
 		}
 	}
@@ -62,11 +62,23 @@ func main() {
 	logger.Infof("running on http://%s (Press Ctrl+C quit)", url)
 	hub := newHub()
 	go hub.run()
+
+	// web file server for working sample service
 	http.HandleFunc("/", func(w http.ResponseWriter, r *http.Request) {
 		http.ServeFile(w, r, "./sample/"+r.URL.Path[1:])
 	})
+	// /ws endpoint is same with /signaling for compatibility
 	http.HandleFunc("/ws", func(w http.ResponseWriter, r *http.Request) {
-		wsHandler(hub, w, r)
+		signalingHandler(hub, w, r)
 	})
-	logger.Fatal(http.ListenAndServe(url, nil))
+	http.HandleFunc("/signaling", func(w http.ResponseWriter, r *http.Request) {
+		signalingHandler(hub, w, r)
+	})
+	// timeout is 10 sec
+	timeout := 10 * time.Second
+	server := &http.Server{Addr: url, Handler: nil, ReadHeaderTimeout: timeout}
+	err := server.ListenAndServe()
+	if err != nil {
+		logger.Fatal(err)
+	}
 }